+++ src/lib/components/organisms/GlobalStats.svelte
@@ -11,6 +11,7 @@
 		powerUpEffectMultiplier,
 		powerUpInterval,
 		xpGainMultiplier,
+		totalBonusPhotonsClicked,
 		totalClicks,
 		totalXP,
 		playerLevel,
@@ -46,12 +47,13 @@
 				<div class="flex-1 flex flex-col gap-6">
 					<h3 class="border-b border-accent/50 pb-2 text-lg font-bold text-accent">Player Stats</h3>
 					<div class="flex flex-col gap-2 xl:gap-4">
-						<StatItem label="Total Clicks" value={formatNumber($totalClicks)} />
+						<StatItem label="Total Clicks" value={formatNumber($totalClicks, 0)} />
+						<StatItem label="Bonus Photons Clicked" value={formatNumber($totalBonusPhotonsClicked, 0)} />
 						<StatItem label="Total XP" value={formatNumber($totalXP)} />
 						<StatItem label="Player Level" value={$playerLevel} />
 						<StatItem label="Atoms per Second" value={formatNumber($atomsPerSecond)} />
 						<StatItem label="Click Power" value={formatNumber($clickPower)} />
-						<StatItem label="Auto Clicks per Second" value={formatNumber($autoClicksPerSecond)} />
+						<StatItem label="Auto Clicks per Second" value={formatNumber($autoClicksPerSecond, 0)} />
 					</div>
 				</div>